refactor: Make component
