+++ feature/home/src/main/java/com/teamoffroad/feature/home/HomeScreen.kt
@@ -56,6 +56,13 @@ import com.teamoffroad.core.designsystem.theme.OpticianSansRegular
 import com.teamoffroad.core.designsystem.theme.PretendardBold
 import com.teamoffroad.core.designsystem.theme.Sub4
 import com.teamoffroad.core.designsystem.theme.White
+import com.teamoffroad.feature.home.presentation.component.CharacterImage
+import com.teamoffroad.feature.home.presentation.component.CharacterNameText
+import com.teamoffroad.feature.home.presentation.component.CloseCompleteRequest
+import com.teamoffroad.feature.home.presentation.component.EmblemNameText
+import com.teamoffroad.feature.home.presentation.component.HomeIcons
+import com.teamoffroad.feature.home.presentation.component.NicknameText
+import com.teamoffroad.feature.home.presentation.component.RecentQuest
 import com.teamoffroad.offroad.feature.home.R
 
 @Composable
@@ -101,114 +108,11 @@ private fun UsersAdventuresInformation(modifier: Modifier = Modifier) {
 }
 
 @Composable
-private fun NicknameText(nickname: String) {
-    Text(
-        modifier = Modifier.padding(start = 24.dp, top = 32.dp),
-        style = OffroadTheme.typography.subtitleReg,
-        text = buildAnnotatedString {
-            append(stringResource(id = R.string.explorer))
-            append(" ")
-            withStyle(
-                SpanStyle(
-                    fontFamily = PretendardBold,
-                    fontWeight = FontWeight.Bold
-                )
-            ) {
-                append(nickname)
-            }
-            append("님")
-        }
-    )
-}
-
-@Composable
-private fun CharacterNameText(name: String) {
-    Text(
-        style = OffroadTheme.typography.subtitle2Semibold,
-        text = name,
-        modifier = Modifier
-            .padding(start = 24.dp, top = 12.dp)
-            .background(
-                color = CharacterName,
-                shape = RoundedCornerShape(6.dp)
-            )
-            .border(
-                width = 1.dp,
-                shape = RoundedCornerShape(6.dp),
-                color = Black15
-            )
-            .padding(horizontal = 16.dp)
-            .padding(vertical = 6.dp),
-        color = White
-    )
-}
-
-@Composable
 private fun HomeBackground() {
     Image(
         painter = painterResource(id = R.drawable.img_home_stamp),
         contentDescription = "stamp",
     )
-}
-
-@Composable
-private fun CharacterImage() {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 90.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.img_home_explorer),
-            contentDescription = "explorer"
-        )
-    }
-}
-
-@Composable
-private fun HomeIcons() {
-    Box(
-        contentAlignment = Alignment.TopEnd,
-        modifier = Modifier.padding(top = 64.dp, end = 28.dp)
-    ) {
-        Column {
-            Image(
-                painter = painterResource(id = R.drawable.ic_home_download),
-                contentDescription = "explorer",
-            )
-            Image(
-                painter = painterResource(id = R.drawable.ic_home_upload),
-                contentDescription = "upload",
-            )
-            Image(
-                painter = painterResource(id = R.drawable.ic_home_change),
-                contentDescription = "change",
-            )
-        }
-    }
-}
-
-@Composable
-private fun EmblemNameText(modifier: Modifier = Modifier) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 24.dp)
-    ) {
-        Box(
-            modifier = modifier.fillMaxWidth(),
-            contentAlignment = Alignment.CenterEnd
-        ) {
-            PopupTagActive(text = "초보 모험가")
-            Image(
-                painter = painterResource(id = R.drawable.ic_home_change_title),
-                contentDescription = "change title",
-                modifier = Modifier.padding(end = 20.dp)
-            )
-        }
-    }
-
 }
 
 @Composable
@@ -235,157 +139,6 @@ private fun UsersQuestInformation() {
             )
         )
         Spacer(modifier = Modifier.padding(end = 24.dp))
-    }
-}
-
-@Composable
-private fun RecentQuest(modifier: Modifier = Modifier, data: HomeProgressBarData) {
-    Surface(
-        color = Contents1,
-        modifier = modifier.clip(shape = RoundedCornerShape(10.dp))
-    ) {
-        Column {
-            Row {
-                ContentsTitle(data.title)
-                Spacer(modifier = Modifier.padding(start = 4.dp))
-                Image(
-                    painter = painterResource(id = R.drawable.img_home_recent_quest),
-                    contentDescription = "recent quest",
-                    modifier = Modifier
-                        .padding(top = 16.dp)
-                )
-            }
-            Spacer(modifier = Modifier.padding(top = 14.dp))
-            CircleProgressBar()
-            Spacer(modifier = Modifier.padding(top = 12.dp))
-            ContentsLocation(data.location)
-            Spacer(modifier = Modifier.padding(bottom = 12.dp))
-        }
-    }
-}
-
-@Composable
-private fun CloseCompleteRequest(modifier: Modifier = Modifier, data: HomeProgressBarData) {
-    Surface(
-        color = Contents2,
-        modifier = modifier
-            .clip(shape = RoundedCornerShape(10.dp))
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Row {
-                ContentsTitle(title = data.title)
-                Spacer(modifier = Modifier.padding(start = 4.dp))
-                Image(
-                    painter = painterResource(id = R.drawable.img_home_close_complete),
-                    contentDescription = "recent quest",
-                    modifier = Modifier
-                        .padding(top = 16.dp)
-                )
-            }
-            Spacer(modifier = Modifier.padding(top = 16.dp))
-            LinearProgressBar(data)
-            Spacer(modifier = Modifier.padding(top = 20.dp))
-            ContentsLocation(data.location)
-            Spacer(modifier = Modifier.padding(bottom = 12.dp))
-        }
-    }
-
-}
-
-@Composable
-private fun CircleProgressBar() {
-    var recentQuestProgress by remember { mutableStateOf(0.8f) }
-
-    ConstraintLayout(
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        val (main, text) = createRefs()
-
-        CircularProgressIndicator(
-            progress = { recentQuestProgress },
-            color = Contents1GraphMain,
-            trackColor = Contents1GraphSub,
-            strokeWidth = 8.dp,
-            strokeCap = StrokeCap.Round,
-            modifier = Modifier
-                .size(82.dp)
-                .constrainAs(main) {
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        )
-
-        Text(
-            text = "3/4",
-            color = Main1,
-            style = OffroadTheme.typography.bothRecentNum,
-            modifier = Modifier
-                .constrainAs(text) {
-                    start.linkTo(main.start)
-                    end.linkTo(main.end)
-                    top.linkTo(main.top)
-                    bottom.linkTo(main.bottom)
-                }
-        )
-    }
-}
-
-@Composable
-private fun LinearProgressBar(data: HomeProgressBarData) {
-    ConstraintLayout {
-        val text = createRef()
-        Text(
-            textAlign = TextAlign.Center,
-            text = buildAnnotatedString {
-                withStyle(
-                    SpanStyle(
-                        fontFamily = OpticianSansRegular,
-                        fontWeight = FontWeight.Normal,
-                        color = Sub4,
-                        fontSize = 62.sp
-                    )
-                ) {
-                    append(data.amount.toString())
-                }
-                withStyle(
-                    SpanStyle(
-                        fontFamily = OpticianSansRegular,
-                        fontWeight = FontWeight.Normal,
-                        fontSize = 30.sp,
-                        color = Black25
-                    )
-                ) {
-                    append("/")
-                    append(data.total.toString())
-                }
-            },
-            modifier = Modifier
-                .fillMaxWidth()
-                .constrainAs(text) {
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        )
-    }
-    Spacer(modifier = Modifier.padding(top = 8.dp))
-    ConstraintLayout {
-        val progress = createRef()
-        var closeCompleteProgress by remember { mutableStateOf(0.8f) }
-
-        LinearProgressIndicator(
-            progress = { closeCompleteProgress },
-            trackColor = Black25,
-            color = Sub4,
-            strokeCap = StrokeCap.Round,
-            modifier = Modifier
-                .height(8.dp)
-                .padding(end = 10.dp)
-                .constrainAs(progress) {
-                    start.linkTo(parent.start, margin = (-3).dp)
-                }
-        )
     }
 }
