+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/NicknameText.kt
@@ -1,0 +1,36 @@
+package com.teamoffroad.feature.home.presentation.component
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.unit.dp
+import com.teamoffroad.core.designsystem.theme.OffroadTheme
+import com.teamoffroad.core.designsystem.theme.PretendardBold
+import com.teamoffroad.offroad.feature.home.R
+
+@Composable
+fun NicknameText(nickname: String) {
+    Text(
+        modifier = Modifier.padding(start = 24.dp, top = 32.dp),
+        style = OffroadTheme.typography.subtitleReg,
+        text = buildAnnotatedString {
+            append(stringResource(id = R.string.explorer))
+            append(" ")
+            withStyle(
+                SpanStyle(
+                    fontFamily = PretendardBold,
+                    fontWeight = FontWeight.Bold
+                )
+            ) {
+                append(nickname)
+            }
+            append("님")
+        }
+    )
+}
